+++ adapters/Facebook/FacebookAdapter/GADFBInterstitialAd.m
@@ -79,6 +79,7 @@
   }
 
   _interstitialAd.delegate = _adapterDelegate;
+  [FBAdSettings setMediationService:[NSString stringWithFormat:@"ADMOB_%@", [GADRequest sdkVersion]]];
   [_interstitialAd loadAd];
 }